+++ src/features/comment/ui/CommentItem.tsx
@@ -4,7 +4,7 @@ import { CommentRow } from "./CommentRow";
 import { Comment } from "../../../entities/comment/model/types";
 import { commentsAtom } from "../../../app/store/atom";
 import { useAtom } from "jotai";
-import { useComment } from "../../../entities/comment/lib/useComment";
+import { useComment } from "../../../entities/comment/hook/useComment";
 
 export const CommentItem: React.FC = () => {
   const [comments] = useAtom(commentsAtom);
